requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql
requests
